All files for aeroelastic loop (first guess)
First eigenstructure.m must be run to find eigenproperties of the plate which is the model for the wing. Loop.m plots for each time step the displacement of a point of the tip of the wing. For each time step the displacement of the wing is calculated by find_w.m (by integration) with lift forces found from force_calc.m and displacement is extrapolated onto the aerodynamic grid by geometry_new.m. For each time step, find_w.m takes as input the initial conditions taken from the preceding find_w.m routine. Further improvements are surely a correction of the time step and the implementation of the unsteady boundary conditions (when calculating RHS in fast_steady)